+++ src/Stack/Types.hs
@@ -17,3 +17,4 @@ import Stack.Types.Image as X
 import Stack.Types.Build as X
 import Stack.Types.Package as X
 import Stack.Types.Compiler as X
+import Stack.Types.Sig as X