+++ ChangeLog.md
@@ -12,6 +12,8 @@ Other enhancements:
   [#1397](https://github.com/commercialhaskell/stack/issues/1397)
 * Only run benchmarks specified as build targets
   [#1412](https://github.com/commercialhaskell/stack/issues/1412)
+* Add optional GPG signing on `stack upload --sign` or with
+  `stack sig sign ...`
 
 Bug fixes: